CU-862jtuxvq - Filter block. FilterCategory and FilterItem components are additionally created to  improve the code
